requirements.txt
# Core frameworks
google-genai==1.62.0
langchain==0.3.27
langchain-core==0.3.83
langgraph==1.0.1
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.1
langgraph-sdk==0.2.15
stagehand-py==0.3.10
boto3==1.35.76
pygeohash==3.2.2

# Browser automation
playwright==1.58.0

# Utility libraries
pydantic==2.12.5
pydantic-settings==2.12.0
python-dotenv==1.2.1
typing-extensions==4.15.0
loguru==0.7.3
tenacity==9.1.2
httpx==0.28.1

# Data handling
python-dateutil==2.9.0.post0
pycountry==26.2.16

# Research tools
tavily-python==0.7.21
firecrawl==4.16.0
